"action" now passed as a callable to Document.udpate_field.
This makes the code more flexible, and easier to update with new
actions. Getting at the update_actions is now a little more obscure --
there aren't as strong pointers in the code to the place where they're
held. And if you mess with the update_actions dict, you can break stuff.

But I think that it is better than the last version of the
code. Possibly when @evansde77's pluggage library is finished and
checked in, we can implement updatea_actions as a plugin.