Fix the Cloudant ctor to handle session headers
- Change implementation to use the X-Cloudant-User header only if the
x_cloudant_user kwarg is supplied along with the url kwarg
- The Cloudant service url is now constructed by first looking for the
account kwarg.  If found then the url is constructed using it.  If not
found then we look for the url and if found we use it as the Cloudant
service url.  In the case where a url kwarg is found we also then check
for the x_cloudant_user kwarg.  If that is found then we add the
X-Cloudant-User to the session object headers.
- Raise an exception if the Cloudant service url is empty after all the
above processing.
- Update tests with account/url/x_cloudant_user kwargs where appropriate