Refactor pylint test
- Remove the assertion based on the STD_ERR.  The new release of pylint
chose to issue informational deprection warnings to STD_ERR which was
triggering the test to fail unnecessarily.  There is no need to check
the STD_ERR for a pylint run because if a problem is encountered the
test will still fail at the final assertion made.  So in a sense the
STD_ERR assertion check is redundant.