Added skip_if_iam to some tests that rely on a legacy password
Some of the bluemix initialization tests make a connection to the service to test the init worked correctly. They use a username and password combination so can't work with an IAM API key without additional switching. While they pass in cases where both legacy creds and IAM creds are supplied, if only IAM creds are supplied they fail so should be disabled for IAM.