Disable false-positive linter error
The build has been failing recently with a false positive linter error.
Pylint was complaining that a method is not callable, even though
the method is defined right on the following line.

As I could not find either the reason pylint was failing or a
solution that resolves this issue, this commit disables the check for
that specific line pylint was failing on.