Override dict's get method
This allows the get method to be able to work with remote databases. Currently calling get only returns a doc if it is in the cache. The original behavior is preserved by use of the `remote` parameter, consistent with other overridden methods such as `keys()`, therefore this should be a non-breaking change.